updates committeds controller

+++ app/controllers/committeds_controller.rb
@@ -1,19 +1,34 @@
 class CommittedsController < ApplicationController
+  before_action :authenticate_user
+
   def index
-    committeds = Committed.all
+    committeds = current_user.committeds
     render json: committeds
   end
 
+  #   committeds_id = params[:id]
+  #   committeds = Committed.find(committeds_id)
+  # OR
+  # committeds = current_user.committeds
+  # OR
+  # committeds = Committed.where(user_id: current_user.id)
+
   def show
-    committeds_id = params[:id]
-    committeds = Committed.find(committeds_id)
-    render json: committeds
+    committeds = Committed.find(params[:id])
+    if current_user.id == committeds.user_id
+      render json: committeds
+    else
+      render json: { message: "Unauthorized: Log into your account!" }, status: :unprocessable_entity
+    end
   end
+
+  # committeds_id = params[:id]
+  # committeds = Committed.find(committeds_id)
 
   def create
     committeds = Committed.new(
       park_id: params[:park_id],
-      user_id: params[:user_id],
+      user_id: current_user.id,
       observations: params[:observations],
       needs: params[:needs],
       teamsize: params[:teamsize],
@@ -24,31 +39,41 @@ class CommittedsController < ApplicationController
     if committeds.save
       render json: committeds
     else
-      render json: { errors: committeds.errors.full_messages }, status: :unprocessable_entity
+      render json: { message: "Unauthorized: Log into your account!" }, status: :unprocessable_entity
     end
   end
 
   def update
-    committeds_id = params[:id]
-    committeds = Committed.find_by(id: committeds_id)
-    committeds.observations = params["observations"] || committeds.observations
-    committeds.needs = params["needs"] || committeds.needs
-    committeds.teamsize = params["teamsize"] || committeds.teamsize
-    committeds.tentativedate = params["tentativedate"] || committeds.tentativedate
-    committeds.experience = params["experience"] || committeds.experience
-    committeds.reason = params["reason"] || committeds.reason
+    committeds = Committed.find(params[:id])
+    if current_user.id == committeds.user_id
+      committeds_id = params[:id]
+      committeds = Committed.find_by(id: committeds_id)
+      committeds.observations = params["observations"] || committeds.observations
+      committeds.needs = params["needs"] || committeds.needs
+      committeds.teamsize = params["teamsize"] || committeds.teamsize
+      committeds.tentativedate = params["tentativedate"] || committeds.tentativedate
+      committeds.experience = params["experience"] || committeds.experience
+      committeds.reason = params["reason"] || committeds.reason
 
-    if committeds.save
-      render json: committeds
+      if committeds.save
+        render json: committeds
+      else
+        render json: { errors: committeds.errors.full_messages }, status: :unprocessable_entity
+      end
     else
-      render json: { errors: committeds.errors.full_messages }, status: :unprocessable_entity
+      render json: { message: "Unauthorized: Log into your account!" }, status: :unprocessable_entity
     end
   end
 
   def destroy
-    committeds_id = params[:id]
-    committeds = Committed.find_by(id: committeds_id)
-    committeds.destroy
-    render json: { message: "Your commitment has been deleted!" }
+    committeds = Committed.find(params[:id])
+    if current_user.id == committeds.user_id
+      # committeds_id = params[:id]
+      # committeds = Committed.find_by(id: committeds_id)
+      committeds.destroy
+      render json: { message: "Your commitment has been deleted!" }
+    else
+      render json: { message: "Unauthorized: Log into your account!" }, status: :unprocessable_entity
+    end
   end
 end
